+++ src/main/java/kitchenpos/application/order/OrderService.java
@@ -9,8 +9,8 @@ import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderRepository;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.OrderTable;
-import kitchenpos.domain.order.OrderTableRepository;
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.table.OrderTableRepository;
 import kitchenpos.dto.order.request.ChangeOrderStatusRequest;
 import kitchenpos.dto.order.request.OrderCreateRequest;
 import kitchenpos.dto.order.request.OrderLineItemRequest;
