+++ src/main/java/kitchenpos/domain/order/Order.java
@@ -15,6 +15,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+import kitchenpos.domain.table.OrderTable;
 
 @Entity
 @Table(name = "orders")
