+++ src/test/java/kitchenpos/application/ServiceTest.java
@@ -1,17 +1,17 @@
 package kitchenpos.application;
 
-import kitchenpos.application.group.TableGroupService;
 import kitchenpos.application.menu.MenuGroupService;
 import kitchenpos.application.menu.MenuService;
 import kitchenpos.application.order.OrderService;
-import kitchenpos.application.order.TableService;
 import kitchenpos.application.product.ProductService;
-import kitchenpos.domain.group.TableGroupRepository;
+import kitchenpos.application.table.TableGroupService;
+import kitchenpos.application.table.TableService;
 import kitchenpos.domain.menu.MenuGroupRepository;
 import kitchenpos.domain.menu.MenuRepository;
 import kitchenpos.domain.order.OrderRepository;
-import kitchenpos.domain.order.OrderTableRepository;
 import kitchenpos.domain.product.ProductRepository;
+import kitchenpos.domain.table.OrderTableRepository;
+import kitchenpos.domain.table.TableGroupRepository;
 import org.assertj.core.data.Percentage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
