+++ src/test/java/kitchenpos/application/table/TableServiceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.order;
+package kitchenpos.application.table;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -7,18 +7,18 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import java.time.LocalDateTime;
 import java.util.List;
 import kitchenpos.application.ServiceTest;
-import kitchenpos.domain.group.TableGroup;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderLineItems;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.OrderTable;
-import kitchenpos.domain.order.OrderTables;
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.table.OrderTables;
+import kitchenpos.domain.table.TableGroup;
 import kitchenpos.dto.order.request.ChangeGuestNumberRequest;
 import kitchenpos.dto.order.request.EmptyOrderTableRequest;
-import kitchenpos.dto.order.request.OrderTableCreateRequest;
-import kitchenpos.dto.order.response.OrderTableResponse;
+import kitchenpos.dto.table.request.OrderTableCreateRequest;
+import kitchenpos.dto.table.response.OrderTableResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
